Switch to using signed integers for Type::bits and Type::width for backward compatibility.
Add -Wsign-compare warning flag as we should have this on everywhere if it is on anywhere.

+++ src/IROperator.h
@@ -156,7 +156,7 @@ EXPORT Expr halide_erf(Expr a);
 
 /** Raise an expression to an integer power by repeatedly multiplying
  * it by itself. */
-EXPORT Expr raise_to_integer_power(Expr a, int b);
+EXPORT Expr raise_to_integer_power(Expr a, int64_t b);
 
 
 }
